+++ scripts/cat2concat.pl
@@ -47,25 +47,26 @@ MAIN:
 		my @F=split /\t/;
 		if($F[8] eq "SM")
 		{
-			my ($GT,$DP,$AF,$ANNOTATION);
+			my ($GT,$DP,$AD,$AF,$ANNOTATION);
 
 			#GT=0/1;DP=60;AF=0.051;DLOOP:CR=1.056772;CP=10.98
 
 			$ANNOTATION="SM=$F[-1]";
-			($F[7]=~/(.*)GT=(.+?);DP=(\d+);AF=(\d+\.\d+)(.*)/ or $F[7]=~/(.*)GT=(.+?);DP=(\d+);AF=(\d+)(.*)/) or die "ERROR: $_";
+			($F[7]=~/(.*)GT=(.+?);DP=(\d+);AD=(\d+);AF=(\d+\.\d+)(.*)/ or $F[7]=~/(.*)GT=(.+?);DP=(\d+);AD=(\d+);AF=(\d+)(.*)/) or die "ERROR: $_";
 			{
 				$GT=$2;
 				$DP=$3;
-				$AF=$4;
+				$AD=$4;
+				$AF=$5;
 
 				$ANNOTATION.=";$1" if($1); 
-				$ANNOTATION.="$5"  if($5);
+				$ANNOTATION.="$6"  if($6);
 				$ANNOTATION=~s/;;/;/;
 			}
 
 			$F[7]=$ANNOTATION;
-			$F[8]="GT:DP:AF";
-			$F[9]="$GT:$DP:$AF";
+			$F[8]="GT:DP:AD:AF";
+			$F[9]="$GT:$DP:$AD:$AF";
 		}
 
 		print join "\t",@F;